+++ docs/CHANGELOG.md
@@ -89,6 +89,7 @@ Version 1.1.0 (Tamarack): Not yet released
 - Add `$Rail Driver Scanner Color` to `af_client.tbl`
 - Add spectate mode options menu
 - Add `dbg_berserk` command for testing berserk entity state
+- Display control binds in spectate mode help test
 
 [@natarii](https://github.com/natarii)
 - Add new damage notification hit sound file